+++ lib/index.js
@@ -15,7 +15,7 @@ var httpAdapter = require('./adapters/http');
 PouchDB.adapter('http', httpAdapter);
 PouchDB.adapter('https', httpAdapter);
 
-PouchDB.adapter('idb', require('./adapters/idb'));
+PouchDB.adapter('idb', require('./adapters/idb/idb'));
 PouchDB.adapter('websql', require('./adapters/websql'));
 PouchDB.plugin(require('pouchdb-mapreduce'));
